refactor: use route groups to isolate embed layout from main site
- Root layout: minimal html/body only (no header/footer)
- (main) route group: header, footer, nav for all main pages
- Embed layout: plain wrapper without site chrome
- All main-site pages moved to app/(main)/
- Shared components stay at app/components/
