+++ app/(main)/admin/manufacturers/[id]/edit/page.tsx
