+++ app/(main)/admin/yachts/[id]/edit/page.tsx
