+++ app/(main)/compare/CompareClient.tsx
