+++ app/(main)/favorites/FavoritesClient.tsx
