+++ app/(main)/favorites/page.tsx
