+++ app/(main)/yachts/YachtsClient.tsx
