+++ app/(main)/yachts/[slug]/RelatedArticles.tsx
