+++ app/(main)/yachts/[slug]/SimilarYachts.tsx
