+++ app/(main)/yachts/[slug]/YachtDetailClient.tsx
