+++ app/(main)/yachts/[slug]/page.tsx
